api/requirements.txt
Flask==3.0.3
flask-cors==4.0.1 #this allows cross-origin requests from my app (IPHONE!)
gunicorn==23.0.0
PyJWT==2.8.0